break out ProjectsField
